calculo inventario Q-P-CGI y registro de orden entrante

+++ Models/DetalleOrdenCompra.cs
@@ -4,7 +4,7 @@
     {
         public virtual long nDetalleOrdenCompra { get; set; }
         public virtual long cantidadArticulos { get; set; }
-        public virtual long precioSubTotal { get; set; }
+        public virtual double precioSubTotal { get; set; }
         public virtual OrdenCompra? ordenCompra { get; set; }
 
         public virtual Articulo? articulo { get; set; }
