+++ Models/VentaDetalle.cs
@@ -4,7 +4,7 @@ namespace Proyect_InvOperativa.Models {
     {
         public virtual long cantidad { get; set; }
         public virtual long nDetalleVenta { get; set; }
-        public virtual long subTotalVenta { get; set; }
+        public virtual double subTotalVenta { get; set; }
         public virtual Ventas? venta {  get; set; }
         public virtual Articulo? articulo { get; set; }
